bugfix: add categories instead of overwriting them
Each answer choice is represented by a `UNNotificationAction` (1 to 1). Bundle these actions into an array and then assign the array as a property of a `UNNotificationCategory`. The category identifier must match that of the `UNMutableNotificationContentin` order for the answers to be coupled together with their Notification.

What’s interesting is that the Category and Content are registered separately.

`notificationCenter.setNotificationCategories([category])`
`notificationCenter.add(request)`

Furthermore, there’s no add equivalent for Category. You just can only set the whole thing. So why not extend `UNUserNotificationCenter` and write an `add(category:)` method?